Moved components to sample folder
